agent/requirements.txt
# Agent package requirements
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
requests>=2.28.0
playwright>=1.40.0
beautifulsoup4>=4.11.0
langfuse>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
